Post-review cleanup: document llm_request_boundary, extract boundaryRecovery module, wire boundary diagnostics (Fixes #2306)
Addresses code-review findings against the issue #2306 implementation:

- docs/hooks/api-reference.md: document the llm_request_boundary field in
  BeforeModelOutput.hookSpecificOutput with the HookLLMRequestBoundary
  interface, contract semantics (suffix verbatim-preserved, prefix
  history-semantics, suffix-fit validity, malformed-as-invalid, differential
  fallback when absent) and a minimal JSON example (issue plan doc task).
- Extract differential boundary recovery, projection snapshots, and hook
  boundary resolution from streamRequestHelpers.ts (969 lines, mixed
  responsibilities) into a cohesive boundaryRecovery.ts module; test file
  renamed to boundaryRecovery.test.ts. Pure move, zero behavior change.
- Wire describeBoundary/ProviderContentBoundary (previously dead code) into
  the resolution diagnostics: caller/hook-metadata/differential log lines now
  carry confidence/pendingStartIndex/pendingCount, with behavioral tests
  asserting the observable log output.
- Collapse redundant identical if/else branches in
  resolvePendingFromHookBoundary (the caller honors onInvalidBoundary).
